+++ examples/SimpleReceiver/PinDefinitionsAndMore.h
@@ -111,20 +111,23 @@
 #undef LED_BUILTIN              // No LED available, take the one which is connected to the DAC output
 #define LED_BUILTIN     PIN_PB5 // 4
 
-#  elif defined(__AVR_ATtiny1614__) && defined(PIN_PA1) // For use with megaTinyCore
+#  elif defined(__AVR_ATtiny1604__) && defined(PIN_PA1) // For use with megaTinyCore
 #define IR_RECEIVE_PIN   PIN_PA1 // 8
-#define IR_SEND_PIN      PIN_PA3 // 10
-#define TONE_PIN         PIN_PA5 // 1
-#define APPLICATION_PIN  PIN_PA4 // 0
-
-#  elif defined(__AVR_ATtiny1604__) && defined(PIN_PA6) // For use with megaTinyCore
-#define IR_RECEIVE_PIN   PIN_PA6 // 2 - To be compatible with interrupt example, pin 2 is chosen here.
 #define IR_SEND_PIN      PIN_PA7 // 3
+#define TONE_PIN         PIN_PA3 // 1  TCA0-WO3
 #define APPLICATION_PIN  PIN_PB2 // 5
 
-#define tone(...) void()      // Define as void, since TCB0_INT_vect is also used by tone()
-#define noTone(a) void()
-#define TONE_PIN         42 // Dummy for examples using it
+#  elif defined(__AVR_ATtiny1614__) && defined(PIN_PA1) // For use with megaTinyCore
+#define IR_RECEIVE_PIN   PIN_PA1 // 8
+#define IR_SEND_PIN      PIN_PA5 // 10 TCA0-WO5
+#define TONE_PIN         PIN_PA3 // 1  TCA0-WO3
+#define APPLICATION_PIN  PIN_PA4 // 0
+
+#  elif  defined(__AVR_ATtiny1624__) && defined(PIN_PA6) // For use with megaTinyCore
+#define IR_RECEIVE_PIN   PIN_PA1 // 8
+#define IR_SEND_PIN      PIN_PA5 // 3  TCA0-WO5
+#define TONE_PIN         PIN_PA3 // 1  TCA0-WO3
+#define APPLICATION_PIN  PIN_PB1 // 6
 
 #  elif defined(__AVR_ATmega1284__) || defined(__AVR_ATmega1284P__) \
 || defined(__AVR_ATmega644__) || defined(__AVR_ATmega644P__) \